adapt code after backport
